requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.2.0
tqdm>=4.61.0
matplotlib>=3.3.4
scikit-learn>=0.24.2
seaborn>=0.11.1
